+++ src/vivarium/framework/population/manager.py
@@ -13,7 +13,6 @@ from typing import Any, Callable, Dict, List, NamedTuple, Optional, Tuple, Union
 
 import pandas as pd
 
-from vivarium import Component
 from vivarium.framework.population.exceptions import PopulationError
 from vivarium.framework.population.population_view import PopulationView
 from vivarium.manager import Manager