+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+**TBD - TBD**
+
+  - Update results-related docstrings
+  
 **3.0.1- 08/20/24**
 
  - Create script to find matching dependency branches